+++ app/Livewire/EventosEditMOdal.php
@@ -28,8 +28,16 @@ class EventosEditMOdal extends Component
     public $fecha;
     #[Validate('required|string|max:65535')]
     public $descripcion = '';
+    #[Validate('required|string|max:65535')]
+    public $descripcion_transf = '';
     #[Validate('nullable|file|max:10240')] // Máximo 10MB (en KB)
     public $archivo; // Aquí se mapea el input file
+    #[Validate('required|numeric|min:0')]
+    public $precio = 0;
+    #[Validate('required|numeric|min:0|max:999')]
+    public $aforo = 0;
+    #[Validate('required|boolean')]
+    public $activo = false;
     
     protected $messages = [
         'nombre.required' => 'El nombre es obligatorio.',
@@ -39,8 +47,17 @@ class EventosEditMOdal extends Component
         'fecha.date' => 'La fecha no es válida.',
         'descripcion.required' => 'La descripción es obligatoria.',
         'descripcion.max' => 'Como máximo 65535 caracteres.',
+        'descripcion_transf.required' => 'La descripción es obligatoria.',
+        'descripcion_transf.max' => 'Como máximo 65535 caracteres.',
         'archivo.file' => 'Debe ser un archivo válido.',
         'archivo.max' => 'El archivo no debe superar los 10MB.',
+        'precio.required' => 'El precio es obligatorio',
+        'precio.numeric' => 'El precio debe ser un numero',
+        'precio.min' => 'El precio debe ser cero o mayor',
+        'aforo.required' => 'El aforo es obligatorio',
+        'aforo.numeric' => 'El aforo debe ser un numero',
+        'aforo.min' => 'El aforo debe ser cero o mayor',
+        'aforo.max' => 'El aforo debe ser menor a 1000',
     ];
 
     public function mount(int $evento_id)
@@ -52,11 +69,15 @@ class EventosEditMOdal extends Component
     {
         if ($this->evento_id && $this->evento_id !== $this->ant_evento_id) {
             $this->evento = Evento::
-                select("id", "nombre", "fecha", "descripcion", "ruta_archivo")
+                select('*')
                 ->find($this->evento_id);
             $this->nombre = $this->evento->nombre;
-            $this->fecha = $this->evento->fecha;
+            $this->fecha = $this->evento->fecha ? $this->evento->fecha->format('Y-m-d') : null;
             $this->descripcion = $this->evento->descripcion;
+            $this->descripcion_transf = $this->evento->descripcion_transferencia;
+            $this->precio = $this->evento->precio;
+            $this->aforo = $this->evento->aforo;
+            $this->activo = $this->evento->activo;
             $this->ant_evento_id = $this->evento_id;
         }
     }
@@ -77,6 +98,10 @@ class EventosEditMOdal extends Component
             $this->evento->nombre = $this->nombre;
             $this->evento->fecha = $this->fecha;
             $this->evento->descripcion = $this->descripcion;
+            $this->evento->descripcion_transferencia = $this->descripcion_transf;
+            $this->evento->precio = $this->precio;
+            $this->evento->aforo = $this->aforo;
+            $this->evento->activo = $this->activo;
             $this->evento->save();
             $this->status['success'][] = 'Evento: ' . $this->evento->nombre . ' fue editado.';
         } else {
@@ -102,8 +127,11 @@ class EventosEditMOdal extends Component
         $this->reset('nombre');
         $this->reset('fecha');
         $this->reset('descripcion');
+        $this->reset('descripcion_transf');
         $this->reset('archivo');
         $this->reset('evento');
+        $this->reset('precio');
+        $this->reset('aforo');
         $this->reset('ant_evento_id');
         $this->resetValidation();
         $this->dispatch('closeModal', ['status' => $this->status]);