+++ app/Models/Adicional.php
@@ -8,4 +8,11 @@ use Illuminate\Database\Eloquent\Model;
 class Adicional extends Model
 {
     use HasFactory;
+
+    protected $fillable = [
+        'cantidad',
+        'nombre',
+        'evento_id',
+        'creador_id',
+    ];
 }