+++ database/migrations/2026_08_11_183805_create_eventos_table.php
@@ -16,10 +16,14 @@ return new class extends Migration
             $table->timestamps();
             $table->string('nombre', 50);
             $table->date('fecha');
+            $table->integer('precio');
+            $table->integer('aforo');
             $table->text('descripcion')->nullable();
+            $table->text('descripcion_transferencia')->nullable();
             $table->unsignedBigInteger('creador_id');
             $table->foreign('creador_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('ruta_archivo')->nullable(); // <--- Usar string (VARCHAR 255)
+            $table->boolean('activo')->default(false);  
         });
     }