+++ resources/views/livewire/eventos-edit-m-odal.blade.php
@@ -8,28 +8,53 @@
                 <form wire:submit="update">
                     <div class="modal-body">
                        <div class="form-row">
-                            <div class="form-group col-md-12">
+                            <div class="form-group col-md-12 p-1">
                                 <label for="name">Nombre: </label>
                                 <input type="text" name="nombre" wire:model.blur="nombre"  class="form-control @error('nombre') is-invalid @enderror">
                                 <div>
                                     @error('nombre') <div class="text-danger">{{ $message }}</div> @enderror
                                 </div>
                             </div>
-                            <div class="form-group col-md-6">
+                            <div class="form-group col-md-6 p-1">
                                 <label for="fecha">Fecha: </label>
-                                <input type="date" name="fecha" wire:model.blur="fecha" class="form-control @error('fecha') is-invalid @enderror">
+                                <input  type="date" 
+                                        name="fecha"
+                                        wire:model.blur="fecha"
+                                        value="{{ $fecha ? \Carbon\Carbon::parse($fecha)->format('Y-m-d') : '' }}"
+                                        class="form-control @error('fecha') is-invalid @enderror">
                                 <div>
                                     @error('fecha') <div class="text-danger">{{ $message }}</div> @enderror
                                 </div>
                             </div>
-                            <div class="form-group col-md-12">
+                            <div class="form-group col-md-6 p-1">
+                                <label for="precio">Precio: </label>
+                                <input type="text" name="precio" wire:model.blur="precio"  class="form-control @error('precio') is-invalid @enderror">
+                                <div>
+                                    @error('precio') <div class="text-danger">{{ $message }}</div> @enderror
+                                </div>
+                            </div>
+                            <div class="form-group col-md-6 p-1">
+                                <label for="aforo">Aforo: </label>
+                                <input type="text" name="aforo" wire:model.blur="aforo"  class="form-control @error('aforo') is-invalid @enderror">
+                                <div>
+                                    @error('aforo') <div class="text-danger">{{ $message }}</div> @enderror
+                                </div>
+                            </div>
+                            <div class="form-group col-md-12 p-1">
                                 <label for="descripcion">Descripción: </label>
                                 <textarea name="descripcion" id="" rows="10" wire:model.blur="descripcion" class="form-control @error('descripcion') is-invalid @enderror"></textarea>
                                 <div>
                                     @error('descripcion') <div class="text-danger">{{ $message }}</div> @enderror
                                 </div>
                             </div>
-                            <div class="form-group col-md-12">
+                            <div class="form-group col-md-12 p-1">
+                                <label for="descripcion_transf">Descripción de la Transferencia: </label>
+                                <textarea name="descripcion_transf" id="" rows="10" wire:model.blur="descripcion_transf" class="form-control @error('descripcion_transf') is-invalid @enderror"></textarea>
+                                <div>
+                                    @error('descripcion_transf') <div class="text-danger">{{ $message }}</div> @enderror
+                                </div>
+                            </div>
+                            <div class="form-group col-md-12 p-1">
                                 <label for="archivo">Flyer: </label>
                                 @if ($evento && $evento->ruta_archivo)
                                     <p>Archivo actual: <a href="{{ asset('storage/' . $evento->ruta_archivo) }}" target="_blank">Ver archivo</a></p>
@@ -54,6 +79,15 @@
                                         <div class="invalid-feedback">{{ $message }}</div>
                                     @enderror
                                 @endif
+                            </div>
+                            <div class="form-group col-md-6 p-2">
+                                <input type="checkbox" 
+                                    id="activo" 
+                                    class="form-check-input" 
+                                    wire:model="activo">
+                                <label class="form-check-label fw-bold" for="activo">
+                                    Evento Activo
+                                </label>
                             </div>
                         </div>
                     </div>