+++ resources/views/livewire/footer.blade.php
@@ -1,7 +1,7 @@
 <div>
         <footer class="bg-body-tertiary fixed-bottom text-center p-2">
-            Developed by Palky Inc. | &copy;{{env('APP_NAME')}}.v{{Config::get('constants.APP_VERSION')}} - 2024
-        | Laravel v{{ Illuminate\Foundation\Application::VERSION }} (PHP v{{ PHP_VERSION }})
+            Developed by Palky Inc. | &copy;{{env('APP_NAME')}}.v{{Config::get('constants.APP_VERSION')}} - 2026
+        {{-- | Laravel v{{ Illuminate\Foundation\Application::VERSION }} (PHP v{{ PHP_VERSION }}) --}}
         </footer>
     </div>
 </div>