+++ resources/views/livewire/navbar.blade.php
@@ -6,32 +6,36 @@
 					<span class="navbar-toggler-icon"></span>
 				</button>
 				<div class="collapse navbar-collapse" id="navbarSupportedContent">
-					<ul class="navbar-nav mr-auto">
-						<li class="nav-item">
-							<a class="nav-link" href="/dashboard">Principal</a>
-						</li>
-						<li class="nav-item">
-						<a class="nav-link disabled" href="/">Lector</a>
-						</li>
-						<li class="nav-item dropdown">
-							<a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-								Datos
-							</a>
-							<ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-								<li><a class="dropdown-item" href="/eventos">Eventos</a></li>
-							</ul>
-						</li>
-						<li class="nav-item dropdown">
-							<a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-								Sistema
-							</a>
-							<ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-								<li><a class="dropdown-item" href="/users">Usuarios</a></li>
-								<li><a class="dropdown-item" href="/roles">Roles</a></li>
-								<li><a class="dropdown-item" href="/permissions">Permisos</a></li>
-							</ul>
-						</li>
-					</ul>
+					@hasrole ('Cliente')
+						<p m-5 p-5>Clientes</p>	
+					@else
+						<ul class="navbar-nav mr-auto">
+							<li class="nav-item">
+								<a class="nav-link" href="/dashboard">Principal</a>
+							</li>
+							<li class="nav-item">
+							<a class="nav-link disabled" href="/">Lector</a>
+							</li>
+							<li class="nav-item dropdown">
+								<a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+									Datos
+								</a>
+								<ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+									<li><a class="dropdown-item" href="/eventos">Eventos</a></li>
+								</ul>
+							</li>
+							<li class="nav-item dropdown">
+								<a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+									Sistema
+								</a>
+								<ul class="dropdown-menu" aria-labelledby="navbarDropdown">
+									<li><a class="dropdown-item" href="/users">Usuarios</a></li>
+									<li><a class="dropdown-item" href="/roles">Roles</a></li>
+									<li><a class="dropdown-item" href="/permissions">Permisos</a></li>
+								</ul>
+							</li>
+						</ul>
+					@endif
 				</div>
 				<div class="collapse navbar-collapse">
 					<a class="navbar-brand" href="">