+++ routes/web.php
@@ -8,6 +8,7 @@ use App\Livewire\Roles;
 use App\Livewire\Dashboard;
 use App\Livewire\Clientes;
 use App\Livewire\Eventos;
+use App\Livewire\Adicionales;
 use App\Http\Controllers\UserController;
 use Illuminate\Foundation\Auth\EmailVerificationRequest;
 
@@ -23,6 +24,7 @@ Route::get('/changeViewMode', [UserController::Class, 'changeViewMode'])->middle
 
 # Routes
 Route::get('/eventos', Eventos::Class)->middleware(['auth', 'verified'])->name('eventos');
+Route::get('/adicionales/{evento_id}', Adicionales::Class)->middleware(['auth', 'verified'])->name('adicionales');
 
 Route::view('profile', 'profile')
     ->middleware(['auth'])