+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Added 
 - Se agrego los modelos y migración de Adicional_Cache, Adicional, Evento, Reserva. 
 - ABM de Eventos.
+- Modal de Reservas
 
 ## [0.0.4] - 2026-08-04
 ### Added